CMPLRLLVM-36077: Restore memory manage transformation for 523.xlancbmk (#3718)
on the new pass manager.

Because the analysis of the key classes' member functions for inlining and
non-inlining is done much earler now, the code is not as completely
optimized as it was when the analysis was done in the inliner itself.

In particular, lifetime intrinsics are still there and the functions can
be slightly larger. I have adjusted the code in MemManageInfoImpl.h to
account for this.

The LIT test includes LLVM bitcode versions of the key classes' member
functions at the time they are now analyzed in the new pass manager.