[NFC] Remove vector-variants annotation from some tests (#332)
It doesn't affect those tests and the semantics of the attribute doesn't make
sense after vectorization has happenned for the rest of the pipeline.

In a later patch I'm planning to add IRVerifier checks to ensure that number of
parameters for the functions matches that of the annotation. The tests updated
here wouldn't satisfy such a check.