[VPlan][HIR] Fix merged CFG-based codegen to support reduction init/final hoisting. (#3692)
When reduction init/final are hoisted out of perfect loop nests
then the reduction temp should be appropriately marked as live-in
for all loops upto the reduction hoist loop. For merged CFG-based
CG loops are attached to parents just-in-time and not ahead-of-time.
This patch updates the mechanism to use OrigLoop instead of
MainLoop to resolve the issue.