[Paropt offload] Fix device compilation bug causing comdat verifier error (#3717)
CMPLRLLVM-34830

In device compilation, if a function F has TARGET regions,
after the TARGET regions are outlined into kernel functions
F is deleted (if it is not marked declare target). However,
if F is a template virtual function then we may hit the
verifier assertion: "Declaration may not be in a Comdat!".

The fix is to remove the comdat from the declaration.