[CMPLRLLVM-35919][1/3] Rename -intel-loop-optreport* options to -intel-opt-report* (#3710)
This series of changes removes "loop" from the names of these internal
options to make them more accurate, since the optimization reports that
they control are not only loop reports but may include OpenMP regions
too. Similar changes have already been made to remove "loop" from the
names of opt-report interfaces, passes, and metadata. Also, this change
expands "optreport" in these options to "opt-report" in order to make
the option names more consistent with other internal options and with
the external -qopt-report* options.

This change is the first of three in order to change these option names
without disrupting driver usability. It adds new -intel-opt-report*
options but keeps the old -intel-loop-optreport* options as aliases. It
will be followed by a change to the driver to change the internal
options used by -qopt-report and then another change to remove the old
-intel-loop-optreport* options once the driver has been updated.