[DTrans - OpaquePtr] Fix for PtrTypeAnalyzer missing identification of some byte-flattened GEPs (#3757)
* [DTrans - OpaquePtr] Fix for PtrTypeAnalyzer missing identification of some byte-flattened GEPs

During testing of some IR being produced with compiler when opaque pointer are in use,
found some cases that triggered DTrans safety issues because the GEP was not identified
as an byte-flattened GEP. Fixed that. Also, don't track a pointer as aliasing an i8*
type if the i8* type is simply for computing a byte-flattened GEP that is identified
as the adress of a specific field, as using i8* for a byte-flattened GEP computation
does not provide useful information for the safety analyzer. i8* will still be added
as a usage type if the GEP cannot be resolved to a specific field.
Updated pointer type collection for store instruction to add a usage type to the types
tracked for the pointer operand for some cases.

* Update test case comments