[Paropt] Fix fp on scalar vars on target, all vars on teams under switches. (#3720)
* [Paropt] Fix fp on scalar vars on target, all vars on teams under switches.

We are currently making non-WI-local firstprivate vars as WI-local on target,
which may not alwas be correct.

Also, we were ignoring firstprivate clause on teams construct
altogether, even though we were handling private. Instead, the items
were being firstprivatized as WI-local on the outer target construct.

For the target construct, the following options can be used:

```
-vpo-paropt-target-non-wilocal-fp-alloc-mode=
   rtl : let libomptarget handle allocation/initialization
   module: allocate in addrspace 1, guard init with master-thread-check
   wilocal: current implementation. treat them as wilocal (default)
```

For the teams construct, the following option can be used to turn
firstprivatization on/off:
```
  -vpo-paropt-handle-firstprivate-on-teams=true/false (default false)
```

For distribut construct, we were unconditionally allocating
firstprivate variables in the stack (private address space). The
following flag can be used to allocate them in addrspace(3) (shared
within workitems in a team):
```
  -vpo-paropt-target-make-distribute-fp-wilocal=true/false (default true)
```

The defaults are set to make this change as an NFC commit. There are
stability issues in FFE that were hidden by the current behavior. We
need for them to be resolved. Also, we may need to work on and improve
the `-vpo-paropt-optimize-data-sharing` pass to counter the peprformance
impact in most common use-cases.

We may consider keeping these flags as-is until we have named-barrier
support, which would avoid having individual master-thread
checks/barriers around stores/calls etc.

Another possible middle ground is to flip these flags if we see any
`omp_get_thread_num`/`omp_get_team_num` calls in the region, to
handle common cases where the current approach would be incorrect.

JIRA: CMPLRLLVM-36097

* Add an assert, minor comment changes.