CMPLRLLVM-32716: Add a zero trip check for unroller/vectorizer's remainder loop. (#3751)
* CMPLRLLVM-32716: Add a zero trip check for unroller/vectorizer's remainder loop.

Some countable loops can have a trip count which is equal to range size of IV
type which is 1 more than the max unsigned val of the type. Special ZTT needs to
be generated in this case for unroller/vectorizer's remainder loop so it executes
the entire iteration range. See comments in HLLoop.h/cpp for more details.