[DTRANS] CMPLRLLVM-35921: Base infrastructure for arrays with constant entries in DTrans and opaque pointers (#3793)
* [DTRANS] CMPLRLLVM-35921: Base infrastructure for arrays with constant entries in DTrans and opaque pointers

This is the basic infrastructure for collecting the information in structures
with fields that are arrays with constant entries. The idea is to check when
a GEP is accessing a field of a structure, and that instruction is followed
by another GEP that access the array entry. Then it will be used by a load
or a store. We are going to keep track of what is being stored, and if
something changes then set the constant value as nullptr.

This is just the basic infrastructure to run some basic test cases. This
work is expected to be expanded once we reach to the actual use case.

* Fix for Robert's comments

* Fix for Chris and Robert comments

* Fix for Chris comment

* Fix for Clang format

* Fix for literal structures

* Fix for sorting literal structures using the printed form