review/sguggill/naive if elim (#3841)
* Remove naive version of block terminator emission

HIR does not allow gotos between then and else children of an HLIf. Our merged
CFG representation uses such gotos and this causes issues when we try to place
then and else children properly. This was addressed earlier by using naive
HLIf generation for the merged CFG path. The changes in this patch eliminate
naive terminator emission interface and update the version that tries to
emit proper HLIfs when possible to do naive emission for ifs outside the
top VPLoop which is where we see the issue with then/else cross linking.

* Fix failure in updated LIT test.