[SLP][NFCI] make LeafData class thinner (#3868)
The goal of the patch is to shrink down size of LeafData class trying to eliminate
all the elements that are not essential for operands reordering. Because we still
reorder operands directly on LLVM IR we need to use backup data. That part was pulled
out of LeafData class and organized as separate storage inside MultiNode class accessed
via dedicated methods to save and lookup the data. Vectorizer methods for saving data
and restoring IR was changed to use this new mechanism.

* few style fixes