[Driver] update object and archive unbundling to allow for SPIRV based obj (#3930)
When unbundling archives and objects, perform an additional
spirv-to-ir-wrapper pass against the extracted device files to check
to see if they are SPIR-V based.  If so, convert them to IR to be
processed down the line.

Additionally, to support this flow, all archives being unbundled will
be unbundled with 'aoo' so we can individually scrutinize the objects
from the archive.  This is a precursor to enabling the ability to filter
out any non-IR (unknown) entities from the device objects/archives.