[DTrans - OpaquePtr] DTrans safety analyzer to handle a case of element zero access in nested type (#3890)
This fixes an safety condition that Robert reported about pointer information
containing multiple element pointees, where one was an element zero of an
inner type, and the other was the element index of the inner type. This avoids
that safety issue. While looking at this, I see that the processing of both
element pointees is redundant, so will follow up with a change that culls the
element pointee set that needs to be processed.