[DTrans - OpaquePtr] Fixes for PtrTypeAnalayzer processing types in 557.xz (#3866)
This fixes the resolution of the element pointee that is targeted
by a byte-flattened GEP that is actually meant to get the address
of a nested type element seen in 557.xz. While investigating this,
I found that some load/store analysis would update the type alias
information tracked  in the declared type and use type lists for
the pointer argument when an element zero type access is made. This
was done when the IR analysis was needing to process bitcasts to
fill in the types, but it does not seem appropriate for the element
zero access to behave this way, so updated to just mark the usage
set, and for loads/stores to be consistent in this.

[CMPLRLLVM-36257]