[Clang][OpenMP] Add an option to allow Intel proprietary opts for device compilation. (#3762)
GPU versions of SVML funcs are effectively a VecClones, so VecClone and VPlan
are required to run to get a proper versions of those functions. Current
device code compilation is prohibited to use these passes with
'-disable-intel-proprietary-opts' cc1 option. But in case of GPU SVML there is
the only alternative place to call VecClone+VPlan - sycl-post-link, but
that would mean procssing the same library functions every compilation which
is incredibly inefficient.

Part of JIRA CMPLRLLVM-28576