CMPLRLLVM-36032: Extend bad casting analysis for 557.xz with opaque pointers (#4029)
* CMPLRLLVM-36032: Extend bad casting analysis for 557.xz with opaque pointers

These changes are needed to get bad casting analysis working on 557.xz with
opaque pointers. They include:
(1) A bug fix in both the typed and opaque pointer version of the bad casting
    analyzer with respect to queueing up PendingStores. This is tested by
    bad-cast-analyzer18.ll.
(2) Special handling for zero element load accesses that only appear in the
    opaque pointer case. This is tested by bad-cast-analyzer19.ll.
(3) Generalization of the testing of dominating of a BasicBlock by a
    conditionally dead basic block or a block with an alloc store. This
    is tested in bad-cast-analyzer20.ll. A wider variety of patterns are
    present when we have opaque pointers, because of the elimination of
    bitcasts, which no longer inhibit movement of loads between basic blocks.
These changes also require the InstructionSimplify.cpp patch from Chris to
conditionalize the required safety checks. But I have verified that it can
be checked in later. They also require Chris's fix (#3994) to remove the
"field address taken call" safety condition. This can also be checked in
independently.

* Fix spelling mistake

* Fixes requested by Chris