[CMPLRLLVM-35919][3/3] Remove old opt-report option spellings (#4021)
This series of changes removes "loop" from the names of these internal
options to make them more accurate, since the optimization reports that
they control are not only loop reports but may include OpenMP regions
too. Similar changes have already been made to remove "loop" from the
names of opt-report interfaces, passes, and metadata. Also, these
changes expand "optreport" in these options to "opt-report" in order to
make the option names more consistent with other internal options and
with the external -qopt-report* options.

Since the first two changes to adopt the -intel-opt-report* spellings in
the optimizer and driver are complete, this final change removes the
aliases for the old -intel-loop-optreport* spellings.

Also, this change fixes a few more uses of the old option spellings that
have slipped back in since #3710 was merged.