[DTrans - OpaquePtr] Create pass to normalize IR for the DTrans safety analyzer (#3994)
* [DTrans - OpaquePtr] Create pass to normalize IR for the DTrans safety analyzer

Introduce a pass to run before passes that need the DTrans safety checks that
puts the IR into a form that can avoid some DTrans safety issues and is more
agreeable to doing DTrans transformations.

In this initial version, this pass will handle one case which is the
generation of GEPs with zero indices when element zero of a structure
is accessed. With opaque pointers, a lot of memory load/store instruction
are using a pointer to the structure to access the 0th element of the
structure even if we suppress the primary instruction simplification that
does this. This occurred in the original DTrans analysis, but is occurring
with greater frequency in the opaque pointer version. These accesses make
it harder for DTrans to operate because now the same pointer may mean a
pointer to an element or a pointer to field within the elements triggering
safety checks, and prevents transformations that want to be performed on
the 0th field because the transformation do not support modifying those cases.

Currently, this only looks at load and store instructions for where to
insert the GEPs. This may be extended in future versions.

* Fix for Prasad's comments

* Fix for Robert's comments