CMPLRLLVM-36026: Relax the conditions to identify if an argument to a function call is a related type (#4057)
* CMPLRLLVM-36026: Relax the conditions to identify if an argument to a function call is a related type

This change set relaxes some conditions to identify if the argument
of a function call passes the related types check (Bad casting for
related types).

First, if there are multipledeclared aliases, we need to find the
alias declared that encapsulates all the aliases through the zero
element. These means that passing any of the types in the declaration
then we will reach the 0 element.

Second, there is a chance that an array type could be in the
zero element access. We need to analyze it.

Third, a pointer to i8 in the aliases declared/used, means that we
are passing the result of a call to new. We can skip this check
since the analyzis for space allocation will take care of it.

* Fix for Robert's comments

* Fix for Chris' comments