[CMake] Install tools to bin/compiler for unified layout builds (#3438)
When INTEL_DEPLOY_UNIFIED_LAYOUT is on deploy LLVM binaries to bin/compiler rather than bin-llvm. This conforms to the upcoming layout redesign in which we shouldn't have special top-level directories, but we can have special subdirectories.

This move will still keep the subdirectory tools out of PATH for customers by default.

Note that nothing turns INTEL_DEPLOY_UNIFIED_LAYOUT on yet, but the expectation is that it will be enabled for xmain-2023 builds in xtoolsup.