cmplrllvm-36376: Fix assert due to null pointer for function address (#4056)
The problem: in window, on host, the mangle name with window ABI is used
for mapping.  But in target the ItaniumABI is used for mapping from
host function.  So in target, the compiler fail to find function address
cause assert.

To fix using ItaniumABI mangled name for mapping instead for all.