[VPlan][HIR] Make the RegDDRef created for VPInvSCEVWrapper base consistent using the ref it originated from (#4079)
* [VPlan][HIR] Make the RegDDRef created for VPInvSCEVWrapper base consistent using the ref it originated from

When creating the base address ref used to compute the number of iterations to dynamically peel for
alignment, we were not making the ref consistent. This caused issues during subsequent vector code
generation. The changes in this patch keep track of the original ref that the base canon expression
originated from during alignment analysis and uses this ref to make the ref for the base address
consistent. The base canon expression for a ref like a[i + n] where a is an i64* looks like
(a + 8 * n). We also needed a slight modification in HIR canon expression checks to allow the same.

* Modify code generation for VPInvSCEVWrapper to generate address of refs of the
form &lp[n1] instead of (lp + 8 * n1) which is not allowed by HIR framework.
This change also reverts the assertion modification in CanonExpr.cpp as it is
not needed any more.

* Fix checks in tests

* Address review comments.