[DTRANS] CMPLRLLVM-36026: Set as Bad casting for related types when a sub instruction is used for allocating a new memory space but there is no dominant aggregate type (#4100)
* [DTRANS] CMPLRLLVM-36026: Set as Bad casting for related types when a sub instrcution is used for allocating a new memory space but there is no dominant aggregate type

This change set checks if the result of a sub instruction is used as
the size for allocating a memory space, but the pointer to the
new space won't have a dominant aggregate type. The reason for not
having the dominant type is because a related type is in the lists of
aliases declared and used. But, the dominant type of the operands used
for the sub operation encapsulates through the zero elemnt all the
types in the aliases declared and used.

Also, the sub operation is guarded with a check that the size is
exactly divisable by the size of the operand's dominat type, is
not equal to zero and less that a limit.

* Fix for Robert's comments

* Fix for Chris' comments

* Fix for Robert's comment