[VPlan][HIR] Fix crash in complete unroller caused by non-constant lower bound in merged CFG path (#4139)
The issue was discovered during extensive alloy testing of peeling and alignment analysis
changes. Consider a loop with a trip count of 3. Planner does a scalar peel of 1 iteration
and chooses VF == 2 for the main loop. Note that we do not need a remainder loop in this
case. Original loop is a constant trip count loop and we were trying to do complete
unroll of the vector loop. However, due to the way the merged CFG is setup, the lower
bound of the vector loop is not a constant and this causes a crash in complete unroller.
The fix checks for no peeling before we do complete unroll.