[NFCI][VecClone] Fix for CMPLRLLVM-9851 (#3556)
* [NFCI][VecClone] Fix for CMPLRLLVM-9851

This patch fixes inconsistencies with how VecClone marks arg memory
as linear/uniform. To be consistent, VecClone will now always create
new memory (i.e., alloca) for the argument, mark appropriately, and
the argument will be stored to this location. A load from this alloca
will be inserted into the loop preheader and any existing loads/stores
to any original allocas for the arg will be replaced with this load.
This makes the argument processing of VecClone consistent between
opt levels, as well.

* [NFCI][VecClone] Small fixes for code review comments