[NFC] CMPLRLLVM-36569: Collect the whole program safe intrinsic name from the actual intrinsic rather tha pass it through a cl opt. (#4165)
The community recently made a couple of changes that requires to
set initialization for the cl opt. If the option is not set as
init, then the information will be set as default. This produced
a problem when setting the name for the whole program safe intrinsic
in the pass that folds the intrinsic. The name was passed through
a cl opt, but then it was set as a default value (an empty string).
The intrinsic folding pass won't fold the intrinsic and will
produce that didn't ran.

This change set stores the name of the intrinsic in a local variable,
and rather than hard code it, we are going to use the function getName
from the Intrinsic namespace in order to collect the actual name
of the whole program safe intrinsic.