[VPlan][TreeConflict] Bug fix for tree conflict running sum (#4185)
This patch fixes a problem where the running sum within the conflict
loop was using fsub/sub instead of fadd/add, even though the reduction
op was fsub/sub. fadd/add needs to be used for sub reduction op cases
so that this value can be subtracted for the final store. Otherwise,
we end up subtracting negative values, which leads to wrong results.