[DTRANS MD] Ensure types referred to by Global variable metadata are emitted in named MD node (#4167)
This extends PR https://github.com/intel-restricted/applications.compilers.llvm-project/pull/3711 to also
include the DTrans metadata that is put on global variables. The TypeFinder by default
only looks at the type (ptr in this case) and the any initializers. This change is
to ensure the type mentioned by the metadata is included in the results for the
types that should be emitted. [CMPLRLLVM-36638]