[DTRANS] CMPLRLLVM-36652: Add -opaque-pointers to all dtrans tests  (#4212)
* [DTRANS] CMPLRLLVM-36652: Add -opaque-pointers to all dtrans tests

Now that the pulldown has happened and opaque pointer support is mostly
in/stable on xmain, add a version of each of the existing dtrans tests
to test directly in opaque-pointers mode.  Note that the metadata of
most of the tests shouldn't change much (if at all)

* Add test coverage to EmptyFPNames, make it work with opaque-ptr

This test apparently didn't really test any of the metadata with
check-lines for some reason, so this patch adds them in. Additionally,
it does so in a way that works with opaque-pointers.

* Fix array check line for LocalConstant

* Fix GlobalLiteralWithInit

This one changes the form in IR slightly, instead of being a literal
struct, this is now just an array of two elements.  IMO, this better
represents the source code anyway, so hopefully this is a
'simplification' of benefit.

* Fix StructWithDeeperPointerTypes

The metadata ends up being a slightly different ordering, so make sure
the test is tolerant of that.