[DTRANS] CMPLRLLVM-36026: Update bad casting for related types to handle return instructions without dominant aggregate types (#4225)
* [DTRANS] CMPLRLLVM-36026: Update bad casting for related types to handle return instructions without dominant aggregate types

This change set updates the check for a return instruction without
a dominant aggregate type, but all the declaration and use aliases
are zero element and/or related types. In this case we are going
to collect the declared type that encloses all the related types
and compare it with the return type. If they are the same then
we are going to set the instrcution and the type as bad casting
for related types.

* Fix for Chris comment