CMPLRLLVM-36553: Inhibit argument promotion when forma/actual args mismatch. (#4271)
I had originally tried fixing this by changing the assert in CallInst::Create
but that caused further problems downstream in the verifier. I had also
thought that only i8 constant 'bool' actuals were being passed down
to i1 bool formals, but it turned out that non-constant actuals were also
involved. So, I decided that the best course of action was to check for
mismatched formals and actuals when we have opaque pointers and avoid the
argument promotion in this case.