Extend the integer parameter when the function isn't local linked (#4184)
We encountered resistance in the community, so I firstly fix it in xmain. Community patch is here: https://reviews.llvm.org/D122963 .
We have to do the extension in the backend since if we remove the zeroext/signext in the FE, it will cause caller won't do the extension.
Adding one option -intel-abi-compatible to the backend, what I want is that we can control all ABI-related behaviors of the backend through this option. Not only for this issue. And also, it can avoid lots of testcase fail. This option will be enabled as default for icx.