[DTRANS MD] No longer limit 'base' emission to polymorphic types (#4353)
See CMPLRLLVM-36729. As reported, we were missing the emission of a
'.base' version of a struct because we mistakenly assumed this was only
valid for 'polymorphic' types, however the example in the test shows
that this is not necessary.  SO this patch should emit the '.base'
version of a function in more cases.