[DTRANS] CMPLRLLVM-36026: Update bad casting for related types to handle when a byte flattened GEP is used to copy data withing a structure' fields (#4358)
* [DTRANS] CMPLRLLVM-36026: Update bad casting for related types to handle when a byte flattened GEP is used to copy data withing a structure' fields

This change set update the analysis for load/store instruction
in DTrans to trace when a byte flattened GEP is used to move
data within the fields of the same structure. The GEP access
multiple fields using a multiplication of the fields size times
a variable to identify which field will be collected and then
it puts the data in an earlier field. This is guarded with the
checks for the language rules. This kind of data movement is
common for std::vectors.

* Fix for Chris' comments

* Fix for Robert's comments