[NFC][VPlan][HIR] Add test to reveal a bug in initial CFG constructed by PlainCFGBuilderHIR. (#4437)
While constructing initial CFG for multi-exit loops, if the label
to jump to from the early exiting block is outside the loop then
we add an edge from exiting block to multi-exit landing pad. This
should not be done in case of multi-exit inner loops where exit
label is still inside the outer loop nest as it can lead to
incorrect landing pads and broken control flow. This patch adds
a test to reveal this bug, fixed in a follow-up PR.

JIRA: CMPLRLLVM-36895
Change-Id: I1f8b3600a6dbed6f1b8f3bde9b229391226a21f0