[VPlan][HIR] Fix incorrect addition of edge from early exiting BB to multi-exit landing pad. (#4438)
While constructing initial CFG for multi-exit loops, if the label
to jump to from the early exiting block is outside the loop then
we add an edge from exiting block to multi-exit landing pad. This
should not be done in case of multi-exit inner loops where exit
label is still inside the outer loop nest as it can lead to
incorrect landing pads and broken control flow. This patch
addresses this issue by updating the check to use entire loop
nest instead of current HLLoop being decomposed.

JIRA : CMPLRLLVM-36895

Change-Id: Ic87ccd751cda74dc0d888c019d99a1dd791f5ea7