[CMake] Use implicit value for INTEL_DEPLOY_UNIFIED_LAYOUT macro (#4451)
In the preprocessor `-DFOO` is equivalent to `-DFOO=1`, but
HandleLLVMOptions.cmake has some issues with the explicit version.

In particular, it tracks definitions as a space-separated list -- but
`add_definitions` and `add_compile_definitions` expect the usual
semicolon-separated lists. An unpleasant side effect of the mismatch is
that the *first* item in the list cannot have an explicit value, or else
subsequent definitions are cosidered part of the first definition's
value. E.g.,

    add_definitions("-DONE -DTWO=2 -DTHREE=3")

yields:

    -DONE -DTWO=2 -DTHREE=3

As expected. However,

    add_definitions("-DONE=1 -DTWO=2 -DTHREE=3")

yields:

    -DONE="1 -DTWO=2 -DTHREE=3"

This results in incorrect definitions for LLVM projects which use
`add_definitions(${LLVM_DEFINITIONS})`.