CMPLRLLVM-36978: Add cold/hot callsites/callees to inlining report. (#4478)
* CMPLRLLVM-36978: Add cold/hot callsites/callees to inlining report.

A key difference between the old and new pass managers is that the inliner
in the new pass manager uses static profiling information to guide the
inlining. So functions and callsites can be considered cold and hot even if
dynamic profiling is not used.

This change set adds support to the inlining report to reflect this. It also
allows hot callees to be recognized by attribute in the same manner as cold
callees are. (This seems to be an oversight that we could push back to the
community.)

I added some LIT tests to llvm/test/Transforms/Intel_InlineReport to check
the new hot/cold inline report messages in both the classic and metadata
inlining reports. I also needed to update one existing InlineReport LIT test
and one in IPO_DeadArgElimination that depends on the exact numbers in the raw
metadata for the metadata inlining report. The new inlining reasons/non-reasons
needed to be added in the appropriate place (not on the end) so that they are
emitted with the appropriate priority order.

* Changes to compiler files

* Replace tabs with spaces