review/erichkeane/opaque ptr tests (#4453)
* [OpaquePtr] Convert some tests to get ready for opaque-ptr

We recently, as a customization set the default opaque-pointer setting
for cc1 to 'off' to get a bunch of tests passing.  However, this is
going to disappear at one point in the near future, as well as cause
us to have failures every time someone in community adds a new test.

This patch adds -opaque-pointer to a large number of tests, and does
what is a fairly mechanical translation of the tests, which went quite
easily once I got into it.  The flag will eventually disappear, but this
will allow us to have these pass now (with the flag default 'off'), and
be a simple removal in the future.

* Fix the lvalue cast tests

* Fix intel tbaa test

* Add 'no-opaque' versions of all the files

* Fix the double-use of no-opaque-pointers and -opaque-pointers on a bunch of the dtrans tests