[DTRANS] CMPLRLLVM-36026: Update "bad alloc size" and "bad ptr manipulation" to check when a subtraction is used with the appropiate size, and "bad ptr manipulator for related types" to handle byte flattened gep (#4426)
* [DTRANS] CMPLRLLVM-36026: Update "bad alloc size" and "bad ptr manipulation" to check when a subtraction is used with the appropiate size, and "bad ptr manipulator for related types" to handle byte flattened gep

This change set fixes 3 simple issues with multiple DTrans
safety violations. First, the dtrans safety "bad alloc size"
couldn't check if the result of a subtraction will be a multiple
of a structure. This was fixed by calling the function that
analyzes the subtraction with the structure size.

Second, the function that checks if a subtraction is multiple
of a structure's size in "bad ptr manipulation" was relaxed
to consider the case where the subtraction is used in a call
to new.

Third, the check for byte flattened GEP was updated to check if
the GEP is accessing the next field in a structure. In this case
we are going to set "bad ptr manipulation for related types". The
utilities for handling related types was updated to handle the
case when this safety violation needs to be reverted as
"bad ptr manipulation".

* Fix Robert's comments

* Fix for Chris' comment