[VPlan][HIR] LIT test to demonstrate an issue in vector code generated for AllZeroCheck instruction (#4508)
LIT test to demonstrate incorrect code generated for AllZeroCheck which is
subsequently used as a branch condition. AllZeroCheck is a uniform instruction.
When generating the vector instruction, all lanes should get the same value.
However, while generating the copy of the widened value, we were marking the
generated instruction as masked, which had the effect of this being a no-op
for inactive lanes. This will cause a problem later when we try to extract lane
0 value and lane 0 happens to be inactive causing wrong branch to be taken.