[XMAIN-WEB] Fix a few lit tests after c79e600 merge. (#4574)
As reported in cmplrllvm-37079.

ldexp-intrin.c was caused by implicit declarations becoming a default
error, so I just disabled it with the flag to get back to our previous
behavior.

The other two I believe the lack of implicit declarations (in cases
where we expect an error) has caused this declaration to no longer
appear (see that in each case, the "did you mean" refers to the error
line!).  This seems like a vast improvement.