[CMPLRLLVM-36647] Avoid placing __svml_* function references in "CG Profile" metadata (#4549)
Calls to the __svml_* functions may be replaced  with a call to a different
function  or scalarized during legalization by the MapIntrinToIml pass, this
can lead to unresolved symbols during linking.
Because the "CG Profile" metadata is just attached to the !llvm.module.flags
metadata node, there is not a reasonable way to locate the information in
the metadata for an affected call that MapIntrinToIml is changing. The only
pass that can reasonably update the metadata is the  GlobalDCE pass, but tha
is a Module level pass so is impractical to run  during the Function pass
pipeline that performs the MapIntrinToIml pass. We don't expect linker function
reordering to be able to do anything with the __svml_ functions, so this should
not impact performance.