[sycl] Prepare FE for upcoming opaque ptr changes
This change adds -opaque-ptr explicitly to CodeGenSYCL tests which
were failing when it was enabled by default.   It then modifies
these tests to handle the equivalent opaque pointer changes in
the emitted IR (ptr instead of i8*, removal of bitcasts etc.)

This change also creates no_opaque_* equivalents for each of these
tests where -no-opaque-ptr is explicitly added to the tests; this
is to continue testing the current behavior.   When opaque pointer
becomes the default and is enabled, the no_opaque_* tests will be
deleted.