[VPlan][HIR] Changes to correctly generate vector instruction for AllZeroCheck
* [VPlan][HIR] Changes to correctly generate vector instruction for AllZeroCheck

The vector copy instruction being generated was being added with a mask causing
this to be a no-op for inactive lanes. AllZeroCheck is a uniform instruction and
all lanes need to get the same value so that later uses(say of lane 0 value) work
correctly. Fixes infinite loop reported in CMPLRLLVM-36680.