CMPLRLLVM-36954 "regF/tr21998@opt_ipo_debug_xG compfail after commit ..." (#4583)
If a routine containing only meta instructions or non-meta instructions
without debug source correlation is processed after a routine with
instructions containing source correlation, then the code can
accidentally create a zero file identifier in the asm case and a crash
in the object file case. Make sure the previous instruction location is
reset to get the correct file id.