Revert "Deploy libffi.so.6 for xmainefi2linux build. (#894)" (#4552)
This reverts commit 1dabe2e6f18997b71bfbda39a520a698ee9c0176, such that
a copy of libffi.so.6 is no longer deployed.

This older copy of libffi was originally deployed becuase
libomptarget.rtl.x86_64.so requires it and some newer distributions
don't package it.

However, the version we deploy (from rdrive) is inadvertently a 32-bit
library. It cannot be used with libomptarget.rtl.x86_64.so.

We have documented that users of some distributions must arrange to
install libffi.so.6 themselves in order to use
libomptarget.rtl.x86_64.so.