[OpaquePtr][OpenMP] Apply the correct addrspace to ident_t* (#4608)
We have to apply addrspace corrections to OpenMP runtime functions but
were explicitly not doing so for IdentTy.  That was done by comparing
pointer types with IdentTy.  With opaque pointers all types are the
same 'ptr' so that does not work.

Now all pointers get the addrspace correction and ident_t* values
must be correctly casted.