[DTRANS] CMPLRLLVM-36026: Add unsafe pointer merge for related types (#4624)
* [DTRANS] CMPLRLLVM-36026: Add unsafe pointer merge for related types

This change set adds the safety flag unsafe pointer merge for
related types. The goal of this safety violation is to check when
we have a PHI node that doesn't have a dominant aggregate type due
to related types, but we know that there is no issue with the
declaration and use aliases.

* Fix for Robert's comment

* Fix for Chris comment, also noticed that BadMemFuncManipulationForRelatedTypes
was in SDArraysWithConstantEntries, when it shouldn't be there.