[SLP] Calculate FMA loss for horizontal reduction (#4614)
There is a latent bug in the SLP vectorizer that was exposed recently when
the cost model was improved to account for downstream optimization of
extractelement instructions inserted by the vectorizer.  With that change,
a computation tree that formerly didn't look profitable now did.
SLP saw an opportunity to create a horizontal reduction (fadds) for a tree
rooted by a bundle of fmuls.  Creating the horizontal reduction after
vectorizing the fmuls was less profitable than leaving the computation alone
and allowing the fmuls and fadds to be contracted into FMAs by the
machine instcombine pass. 
We need to recognize that generating a horizontal reduction
removes the opportunity to form some FMA instructions during the machine
instcombine pass.  This patch adds a cost estimate based on the number of
FMAs that would otherwise be generated.