[OpaquePtr][OpenMP] Begin modifying tests to use OpenMP TYPED clauses (#4667)
Adds -fopenmp-typed-clauses option to opaque pointer enabled tests.

Modifies tests to expect TYPED qualifier. For the most part, no attempt is
made to add the additional 'type' and 'num-elements' arguments.  That would
take an unreasonable amount of time and doesn't help much since in
general these tests were not written to verify types.

Going forward, new tests should include the type and number of elements
if that is important to whatever needs testing.