Disable x86-gather-to-load-permute to act on arrays of size equal 1. (#4691)
Arrays of size 1 are usually used as a placeholders for arrays of a
bigger size kept in the structures so it's incorrect to apply the
optimization to them. Also, it seems very unlikely that someone
will use a gather-like load from array of size 1.
Not disabling the transformation on such arrays leads to run fail
of cpu2017/602,502 during fixing the CMPLRLLVM-24868.