review/skozhukh/masked rem 2 code (#4588)
[VPlan] Disable non-phi and non-select as liveouts for conditional lastprivates.

In some cases optimizations can modify conditional updates into
unconditional once. E.g.
   if (cond)
     ptr += 2
can be transformed into
   off = cond ? 2 : 0;
   ptr += off;
ptr is liveout.

In the first case we have an explicit use in update instruction
and a liveout phi. In the second case we have liveout update where
the header phi is used.

The implicit last private analysis considers the exit instruction as
a valid use of the header phi. But the valid use can be only phi or
select.